Calculate first-order, GLCM, and/or GLRLM radiomic features on the whole 3D lung

This is a wrapper for RIA R package. It calculates first-order, GLCM, and/or GLRLM on the whole 3D lung, left and right lungs separately

RIA_lung(img, mask, sides = c("right", "left"), features = c("fo",
  "glcm", "glrlm"), bins_in = 8, equal_prob = FALSE, distance = 1,
  statistic = "mean(X, na.rm = TRUE)", verbose_in = TRUE)

Arguments

img
CT scan in ANTs image file format

mask
Mask of CT scan in ANTs image file format

sides
Choose to calculate radiomic features on the right and/or left lungs. Note: Right lung = 1, left lung = 2, non-lung = 0

features
Type of radiomic features to calculate. Options: first-order, GLCM, and/or GLRLM

bins_in
Number of bins to discretize image

equal_prob
logical, indicating to cut data into bins with equal relative frequencies. If FALSE, then equal interval bins will be used.

distance
integer, distance between the voxels being compared.

statistic
string, defining the statistic to be calculated on the array of GLCM statistics. By default, statistic is set to "mean", however any function may be provided. The proper syntax is: function(X, attributes). The supplied string must contain a "X", which will be replaced with the array of the GLCM statistics value. Further attributes of the function may also be given. For example, if you wish to calculate the median of all GLCMs calculated in different directions, then it must be supplied as: median(X, na.rm = TRUE).

verbose_in
logical, indicating whether to print detailed information. Most prints can also be suppressed using the suppressMessages function.

Value

list containing the statistical information