Radiomic Calculation on Lung CTs

Calculate individual radiomic features on the whole 3D lung, left and right lungs separately

radiomics_lung(img, mask, sides = c("right", "left"),
  featuresFirst = c("mean", "sd", "skew", "kurtosis", "min", "q1",
  "median", "q3", "max", "energy", "rms", "uniformity", "entropy"),
  featuresSpatial = c("mi", "gc", "fd"))

Arguments

img
CT scan in ANTs image file format

mask
Mask of CT scan in ANTs image file format

sides
Choose to calculate radiomic features on the right and/or left lungs. Note: Right lung = 1, left lung = 2, non-lung = 0

featuresFirst
First level radiomic features to calculate

featuresSpatial
Spatial radiomic features to calculate

Value

Values from selected features for both left and right lungs