Radiomic Calculation on Partitioned Lung

Calculate radiomic features on the partitioned 3D lung

radiomics_partition(
  img,
  mask,
  sides = c("right", "left"),
  featuresFirst = c("mean", "sd", "skew", "kurtosis", "min", "q1", "median", "q3",
    "max", "energy", "rms", "uniformity", "entropy"),
  featuresSpatial = c("mi", "gc", "fd"),
  partition = NULL,
  kernel_size = c(30, 30, 30),
  kernel_stride = c(30, 30, 30),
  threshold = 1000,
  tidy = TRUE
)

Arguments

img
CT scan in ANTs image file format

mask
Mask of CT scan in ANTs image file format

sides
Choose to calculate radiomic features on the right and/or left lungs. Note: Right lung = 1, left lung = 2, non-lung = 0

featuresFirst
First level radiomic features to calculate

featuresSpatial
Spatial radiomic features to calculate

partition
Matrix of x, y, and z coordinates for each partition from partition_lung. If null, partition_lung is called.

kernel_size
(If partition is null) Size of the kernel, in voxel units of width, depth, and height. Must be c(3,3,3) or greater. Default: c(30,30,30)

kernel_stride
(If partition is null) Stride (or spacing) between kernels, in voxel units, for width, depth, and height. If kernel_stride = kernel_size, the partitions are non-overlapping. If stride = c(1,1,1), then each voxel is returned.

threshold
Number of non-missing voxels needed to calculate radiomic features in each partition.

tidy
Logical. If true, outputs a tidy dataframe with results. If false, outputs nested loop.

Value

Values from selected features for both left and right lungs